backend/foodgram/requirements.txt
asgiref==3.5.2
celery==5.2
eventlet
Django==4.1.8
django-filter==2.4.0
django-cors-headers
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
drf-yasg==1.21.3
flower
gunicorn==20.0.4
psycopg2-binary==2.9.4
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
PyJWT==2.1.0
pytz==2021.3
sqlparse==0.3.1
requests==2.26.0
reportlab
python-dotenv
Pillow
flake8
flake8-isort
flake8-return
flake8-broken-line
pep8-naming